💥 Rename `other` module to `misc`
