+++ datar/all.py
@@ -13,7 +13,7 @@ from . import base as _base, dplyr as _dplyr
 from .forcats import *
 from .tibble import *
 from .tidyr import *
-from .other import *
+from .misc import *
 
 _locs.update(
     {
